Refactor cus 04 (#39)
* refactor : 변수와 경로명을 잘못 적어 코드 실행이 안되는 문제 해결

* refactor : 예약하려는 주차장소가 해당 주차장에 없을경우 방어로직 추가, 예약 취소에서 id를 reservationId로 변경
