+++ .gitignore
@@ -37,4 +37,4 @@ out/
 .vscode/
 
 # 로컬 설정 파일 제외
-src/main/resources/application-local.yml
+src/main/resources/application.yml
