+++ src/main/java/com/example/parking/global/initdata/ParkingDataRunner.java
@@ -1,12 +1,7 @@
 package com.example.parking.global.initdata;
 
-import com.example.parking.domain.parking.entity.ParkingLot;
-import com.example.parking.domain.parking.entity.SpotType;
-import com.example.parking.domain.parking.repository.ParkingLotRepository;
-import com.example.parking.domain.parkingspot.entity.ParkingSpot;
+import com.example.parking.domain.parkingLot.repository.ParkingLotRepository;
 import com.example.parking.domain.parkingspot.repository.ParkingSpotRepository;
-import com.example.parking.domain.reservation.entity.Reservation;
-import com.example.parking.domain.reservation.entity.ReservationStatus;
 import com.example.parking.domain.reservation.repository.ReservationRepository;
 import com.example.parking.domain.user.entity.User;
 import com.example.parking.domain.user.entity.UserRole;
@@ -16,9 +11,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
-import java.time.LocalTime;
 
 @Component
 @RequiredArgsConstructor
@@ -44,31 +36,32 @@ public class ParkingDataRunner implements CommandLineRunner {
                 .role(UserRole.USER)
                 .build());
 
-        // 2. 주차장 (ID: 1)
-        ParkingLot parkingLot = parkingLotRepository.save(ParkingLot.builder()
-                .name("강남 공영 주차장")
-                .address("서울 강남구")
-                .totalSpot(50)
-                .price(3000)
-                .operationStartTime(LocalTime.of(9, 0))
-                .operationEndTime(LocalTime.of(22, 0))
-                .build());
+//        // 2. 주차장 (ID: 1)
+//        ParkingLot parkingLot = parkingLotRepository.save(ParkingLot.builder()
+//                .name("강남 공영 주차장")
+//                .address("서울 강남구")
+//                .totalSpot(50)
+//                .price(3000)
+//                .operationStartTime(LocalTime.of(9, 0))
+//                .operationEndTime(LocalTime.of(22, 0))
+//                .build());
 
-        // 3. 주차 자리 (ID: 1)
-        ParkingSpot parkingSpot = parkingSpotRepository.save(ParkingSpot.builder()
-                .parkingLot(parkingLot)
-                .number("A-01")
-                .type(SpotType.SMALL)
-                .build());
-
-        // 4. 예약 (ID: 1) - status 필드 추가
-        reservationRepository.save(Reservation.builder()
-                .user(user)
-                .parkingLot(parkingLot)
-                .parkingSpot(parkingSpot)
-                .startTime(LocalDateTime.now().plusHours(1))
-                .endTime(LocalDateTime.now().plusHours(3))
-                .status(ReservationStatus.CONFIRMED) // 🔥 신버전: 초기 상태 설정 (NULL 에러 방지)
-                .build());
+//        // 3. 주차 자리 (ID: 1)
+//        ParkingSpot parkingSpot = parkingSpotRepository.save(ParkingSpot.builder()
+//                .parkingLot(parkingLot)
+//                .number("A-01")
+//                .type(SpotType.SMALL)
+//                .build());
+//
+//        // 4. 예약 (ID: 1) - status 필드 추가
+//        reservationRepository.save(Reservation.builder()
+//                .user(user)
+//                .parkingLot(parkingLot)
+//                .parkingSpot(parkingSpot)
+//                .startTime(LocalDateTime.now().plusHours(1))
+//                .endTime(LocalDateTime.now().plusHours(3))
+//                .status(ReservationStatus.CONFIRMED) // 🔥 신버전: 초기 상태 설정 (NULL 에러 방지)
+//                .build());
+//    }
     }
 }
