+++ src/test/java/com/example/parking/domain/reservation/service/ReservationServiceTest.java
@@ -1,9 +1,9 @@
 package com.example.parking.domain.reservation.service;
 
-import com.example.parking.domain.parking.entity.ParkingLot;
+import com.example.parking.domain.parkingLot.entity.ParkingLot;
 import com.example.parking.domain.parkingspot.entity.ParkingSpot;
-import com.example.parking.domain.parking.entity.SpotType;
-import com.example.parking.domain.parking.repository.ParkingLotRepository;
+import com.example.parking.domain.parkingLot.entity.SpotType;
+import com.example.parking.domain.parkingLot.repository.ParkingLotRepository;
 import com.example.parking.domain.parkingspot.repository.ParkingSpotRepository;
 import com.example.parking.domain.reservation.dto.ReservationResDto;
 import com.example.parking.domain.reservation.entity.Reservation;
@@ -93,7 +93,7 @@ class ReservationServiceTest {
         // given (주어진 상황: setUp에서 이미 가짜 데이터 1건이 세팅됨)
 
         // when (실행: 재현님이 작성한 서비스 로직을 호출한다)
-        List<ReservationResDto> results = reservationService.getMyReservations(savedUser.getId());
+        List<ReservationResDto> results = reservationService.getMyReservations(savedUser.getId(), null);
 
         // then (검증: 결과가 내 예상과 똑같은지 확인한다)
         assertThat(results).hasSize(1); // 1건이 나와야 함
